Reduce execution time for SAMLQueryProfileTestCase
