+++ pom.xml
@@ -1934,7 +1934,7 @@
 
 
         <!-- Identity Inbound Versions   -->
-        <identity.inbound.auth.saml.version>5.7.6-SNAPSHOT</identity.inbound.auth.saml.version>
+        <identity.inbound.auth.saml.version>5.7.7</identity.inbound.auth.saml.version>
         <identity.inbound.auth.oauth.version>6.2.68</identity.inbound.auth.oauth.version>
         <identity.inbound.auth.openid.version>5.4.3</identity.inbound.auth.openid.version>
         <identity.inbound.auth.sts.version>5.4.3</identity.inbound.auth.sts.version>
